Fix mocha tests + path resolve bug
So many levels of errors! Mocha test were failing because
minijasminenode was resolving the path, so that was a real bug. Also,
tests were timing out, so I deactivated the gulp test for the timeout
setting, and since the reporter doesn't support the verbose setting, I
changed the test pass condition, but I'm not proud of my solution...